+++ packages/koot/core/webpack/defaults/isomorphic.server.dev.js
@@ -15,6 +15,7 @@ const factoryConfig = async ({
 
     return {
         mode: "development",
+        devtool: 'source-map',
         target: 'async-node',
         node: {
             __dirname: true