+++ packages/koot/core/webpack/defaults/isomorphic.server.prod.js
@@ -10,6 +10,7 @@ const factoryConfig = async ({
 
     return {
         mode: "development",
+        devtool: 'source-map',
         target: 'async-node',
         node: {
             __dirname: true